fix: use int64_t for timestamps to prevent overflow on long videos
Videos longer than ~13.5 hours crash due to audio timestamp overflow.
At 44.1kHz, int32 overflows after 2,147,483,647 / 44,100 ≈ 48,662 seconds.

Changed to int64_t:
- audio_time, video_time, duration in stream.h
- sync_v_time, sync_a_time, audio_ref_time, audio_samples
- Function parameters in stream_sync.c and stream_audio.c
- Internal duration/position handling in avos_mp*.c

Fixes: nova-video-player/aos-Video#185

+++ Include/avos_mp_priv.h
@@ -26,7 +26,7 @@ typedef struct avos_mp_audio avos_mp_audio_t;
 
 avos_mp_video_t *avos_mp_getvideo(avos_mp_t *mp);
 avos_mp_audio_t *avos_mp_getaudio(avos_mp_t *mp);
-int avos_mp_fillmetadata(avos_mp_t *mp, int type, uint64_t size, ID3_TAG *id3_tag, AV_PROPERTIES *av, const char *mimetype, int duration, int seekable, int pauseable, int decoder); /* return 1 if metadata changed */
+int avos_mp_fillmetadata(avos_mp_t *mp, int type, uint64_t size, ID3_TAG *id3_tag, AV_PROPERTIES *av, const char *mimetype, int64_t duration, int seekable, int pauseable, int decoder); /* return 1 if metadata changed */
 
 void avos_mp_sendevent(avos_mp_t *mp, int what, int arg1, int arg2);
 void avos_mp_sendevent_msg(avos_mp_t *mp, int what, int arg1, int arg2, const char *msg);
